apm0013173-carche/requirements.txt
Django>=1.10,<2
django-cors-headers==1.3.1
configparser==3.5.0
djangorestframework>=3.5.3
drf-tracking
pymysql
Jinja2>2.10
jinja2schema==0.1.4
jsonpath-ng==1.5.2
netaddr==0.8.0
requests
regex